backend/requirements.txt
fastapi
uvicorn
python-multipart
beautifulsoup4
playwright
google-genai
python-dotenv
supabase